+++ src/lyrics_fetcher.h
@@ -26,6 +26,8 @@
 #include <memory>
 #include <string>
 
+#include "song.h"
+
 struct LyricsFetcher
 {
 	typedef std::pair<bool, std::string> Result;
@@ -33,7 +35,7 @@ struct LyricsFetcher
 	virtual ~LyricsFetcher() { }
 
 	virtual const char *name() const = 0;
-	virtual Result fetch(const std::string &artist, const std::string &title);
+	virtual Result fetch(const std::string &artist, const std::string &title, const MPD::Song &song);
 	
 protected:
 	virtual const char *urlTemplate() const = 0;
@@ -57,7 +59,7 @@ std::istream &operator>>(std::istream &is, LyricsFetcher_ &fetcher);
 
 struct GoogleLyricsFetcher : public LyricsFetcher
 {
-	virtual Result fetch(const std::string &artist, const std::string &title);
+	virtual Result fetch(const std::string &artist, const std::string &title, const MPD::Song &song);
 	
 protected:
 	virtual const char *urlTemplate() const { return URL; }
@@ -120,7 +122,7 @@ protected:
 struct InternetLyricsFetcher : public GoogleLyricsFetcher
 {
 	virtual const char *name() const override { return "the Internet"; }
-	virtual Result fetch(const std::string &artist, const std::string &title) override;
+	virtual Result fetch(const std::string &artist, const std::string &title, const MPD::Song &song) override;
 	
 protected:
 	virtual const char *siteKeyword() const override { return nullptr; }
@@ -131,5 +133,17 @@ protected:
 private:
 	std::string URL;
 };
+
+#ifdef HAVE_TAGLIB_H
+struct TagsLyricsFetcher : public LyricsFetcher
+{
+	virtual const char *name() const override { return "tags"; }
+	virtual Result fetch(const std::string &artist, const std::string &title, const MPD::Song &song) override;
+
+protected:
+	virtual const char *urlTemplate() const override { return ""; }
+	virtual const char *regex() const override { return ""; }
+};
+#endif // HAVE_TAGLIB_H
 
 #endif // NCMPCPP_LYRICS_FETCHER_H