+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Add `visualizer_spectrum_smooth_look_legacy_chars` option (enabled by default)
   for potentially improved bottom part of the spectrum visualizer in terminals
   with transparent background.
+* Add support for fetching lyrics from tags.
 
 # ncmpcpp-0.9.2 (2021-01-24)
 * Revert suppression of output of all external commands as that makes e.g album